refactor(portal-next): create tabs view component for catalog
