+++ gravitee-apim-portal-webui-next/src/app/catalog/tabs-view/tabs-view.component.scss
@@ -1,46 +1,46 @@
 /*
- * Copyright (C) 2024 The Gravitee team (http://gravitee.io)
- *
+ * Copyright (C) 2025 The Gravitee team (http://gravitee.io)
+ * 
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- *
+ * 
  *         http://www.apache.org/licenses/LICENSE-2.0
- *
+ * 
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-@use '../../scss/theme';
-
 :host {
   display: flex;
   flex-flow: column;
   gap: 32px;
 }
 
-.welcome-banner {
-  display: flex;
-  flex-flow: column;
-  padding: 16px;
-  gap: 36px;
+.api-list__container {
+  display: grid;
+  gap: 32px;
+  grid-template-columns: repeat(3, minmax(0, 1fr));
+}
 
-  &__content {
+.catalog {
+  &__filters {
     display: flex;
     flex-flow: column;
-    gap: 24px;
-  }
+    gap: 32px;
 
-  &__actions {
-    display: flex;
-    align-items: center;
-    gap: 20px;
+    &__bottom-row {
+      display: flex;
+      justify-content: space-between;
 
-    &__secondary-button {
-      color: theme.$banner-text-color;
-      text-decoration: none;
+      &__filter-labels {
+        display: flex;
+        flex-flow: column;
+        justify-content: center;
+        gap: 8px;
+      }
     }
   }
 }
